Abstract exposure transfer
