+++ back/app/controllers/web_api/v1/confirmations_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class WebApi::V1::ConfirmationsController < ApplicationController
+  include AnonymousExposureTransfer
+
   skip_before_action :authenticate_user, only: %i[confirm_code_unauthenticated]
   skip_after_action :verify_authorized
 
@@ -17,6 +19,7 @@ class WebApi::V1::ConfirmationsController < ApplicationController
 
     if result.success?
       SideFxUserService.new.after_update(user, user)
+      transfer_anonymous_exposures(user)
 
       payload = user.to_token_payload
       auth_token = AuthToken::AuthToken.new payload: payload
